refactor: 중복된 file 코드 통합
