+++ tomcat/src/main/java/org/apache/coyote/http11/util/Resource.java
@@ -1,0 +1,20 @@
+package org.apache.coyote.http11.util;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import org.apache.coyote.http11.servlet.Servlet;
+
+public class Resource {
+    public static String getFile(String fileName){
+        try {
+            final var fileUrl = Servlet.class.getClassLoader().getResource("static/" + fileName);
+            final var fileBytes = Files.readAllBytes(new File(fileUrl.getFile()).toPath());
+            return new String(fileBytes);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        } catch (NullPointerException e){
+            return "";
+        }
+    }
+}
